python/requirements.txt
Pillow  # ==7.1.2
numpy
OpenCV-Python==4.3.0